refactor: move Sentry client helpers to lib/sentry
`apps/web/app` is for App Router routes and route groups; per AGENTS.md
services belong in `apps/web/lib`. The Sentry init logic, its runtime-config
types and the config-script helper are none of those, so they move to
`apps/web/lib/sentry`. `instrumentation-client.ts` stays at the app root --
Next.js resolves that exact path.
